feat: added coffee cog
